[WFLY-17275] Upgrade the jaxrs subsystem model to 4.0 and the schema to 3.0.
https://issues.redhat.com/browse/WFLY-17275
