+++ jaxrs/src/test/java/org/jboss/as/jaxrs/JaxrsSubsystem30TestCase.java
@@ -24,6 +24,7 @@ package org.jboss.as.jaxrs;
 import java.io.IOException;
 
 import org.jboss.as.subsystem.test.AbstractSubsystemBaseTest;
+import org.junit.Test;
 
 /**
  * @author <a href="kabir.khan@jboss.com">Kabir Khan</a>
@@ -31,24 +32,29 @@ import org.jboss.as.subsystem.test.AbstractSubsystemBaseTest;
  * @author <a href="mailto:alessio.soldano@jboss.com>Alessio Soldano</a>
  * @author <a href="rsigal@redhat.com>Ron Sigal</a>
  */
-public class JaxrsSubsystem20MinimalTestCase extends AbstractSubsystemBaseTest {
+public class JaxrsSubsystem30TestCase extends AbstractSubsystemBaseTest {
 
-    public JaxrsSubsystem20MinimalTestCase() {
+    public JaxrsSubsystem30TestCase() {
         super(JaxrsExtension.SUBSYSTEM_NAME, new JaxrsExtension());
     }
 
     @Override
     protected String getSubsystemXml() throws IOException {
-        return readResource("jaxrs-2.0_minimal.xml");
+        return readResource("jaxrs.xml");
     }
 
     @Override
     protected String getSubsystemXsdPath() throws Exception {
-        return "schema/jboss-as-jaxrs_2_0.xsd";
+        return "schema/jboss-as-jaxrs_3_0.xsd";
     }
 
     @Override
     public void testSubsystem() throws Exception {
-        standardSubsystemTest(null, false);
+        standardSubsystemTest(null);
+    }
+
+    @Test
+    public void testExpressions() throws Exception {
+        standardSubsystemTest("jaxrs-expressions.xml");
     }
 }
