+++ jaxrs/src/test/resources/org/jboss/as/jaxrs/jaxrs.xml
@@ -1,0 +1,38 @@
+<subsystem xmlns="urn:jboss:domain:jaxrs:3.0">
+    <jaxrs-2-0-request-matching>true</jaxrs-2-0-request-matching>
+    <resteasy-add-charset>false</resteasy-add-charset>
+    <resteasy-buffer-exception-entity>false</resteasy-buffer-exception-entity>
+    <resteasy-disable-html-sanitizer>true</resteasy-disable-html-sanitizer>
+    <resteasy-disable-providers>
+        <class>a.b</class>
+        <class>c</class>
+    </resteasy-disable-providers>
+    <resteasy-document-expand-entity-references>true</resteasy-document-expand-entity-references>
+    <resteasy-document-secure-disableDTDs>false</resteasy-document-secure-disableDTDs>
+    <resteasy-document-secure-processing-feature>false</resteasy-document-secure-processing-feature>
+    <resteasy-gzip-max-input>777</resteasy-gzip-max-input>
+    <resteasy-jndi-resources>
+       <jndi>java:a/b</jndi>
+       <jndi>java:k.l/mm/nnn</jndi>
+    </resteasy-jndi-resources>
+    <resteasy-language-mappings>
+        <entry key="o">4444</entry>
+        <entry key="pp">5555</entry>
+        <entry key="q">6666</entry>
+    </resteasy-language-mappings>
+    <resteasy-media-type-mappings>
+        <entry key="r">77777</entry>
+    </resteasy-media-type-mappings>
+    <resteasy-media-type-param-mapping>tu</resteasy-media-type-param-mapping>
+    <resteasy-prefer-jackson-over-jsonb>true</resteasy-prefer-jackson-over-jsonb>
+    <resteasy-providers>
+      <class> s </class>
+      <class> t.t </class>
+    </resteasy-providers>
+    <resteasy-rfc7232preconditions>true</resteasy-rfc7232preconditions>
+    <resteasy-role-based-security>true</resteasy-role-based-security>
+    <resteasy-secure-random-max-use>2222</resteasy-secure-random-max-use>
+    <resteasy-use-builtin-providers>false</resteasy-use-builtin-providers>
+    <resteasy-use-container-form-params>true</resteasy-use-container-form-params>
+    <resteasy-wider-request-matching>true</resteasy-wider-request-matching>
+</subsystem>
